requirements.txt
torch
torchvision
transformers
ultralytics
numpy
Pillow
opencv-python-headless
fastapi
uvicorn
gdown
scikit-learn